+++ Sources/DCSettings/Extensions/Color+Codable.swift
@@ -13,12 +13,6 @@ import UIKit
 import AppKit
 #endif
 
-#if compiler(>=6.0)
-extension Color: @retroactive Codable {}
-#else
-extension Color: Codable {}
-#endif
-
 extension Color {
 
     #if canImport(UIKit)
@@ -27,6 +21,30 @@ extension Color {
     typealias NativeColor = NSColor
     #endif
 
+    func dcSettingsEncodedData() throws -> Data {
+        let nativeColor = NativeColor(self)
+        var red: CGFloat = 0
+        var green: CGFloat = 0
+        var blue: CGFloat = 0
+        var alpha: CGFloat = 0
+        nativeColor.getRed(&red, green: &green, blue: &blue, alpha: &alpha)
+
+        let components = DCColorComponents(red: red, green: green, blue: blue, opacity: alpha)
+        return try JSONEncoder().encode(components)
+    }
+
+    init(dcSettingsData data: Data) throws {
+        let components = try JSONDecoder().decode(DCColorComponents.self, from: data)
+        self.init(NativeColor(red: components.red, green: components.green, blue: components.blue, alpha: components.opacity))
+    }
+}
+
+private struct DCColorComponents: Codable {
+    let red: CGFloat
+    let green: CGFloat
+    let blue: CGFloat
+    let opacity: CGFloat
+
     private enum CodingKeys: String, CodingKey {
         case red
         case green
@@ -34,29 +52,26 @@ extension Color {
         case opacity
     }
 
-    public func encode(to encoder: Encoder) throws {
+    init(red: CGFloat, green: CGFloat, blue: CGFloat, opacity: CGFloat) {
+        self.red = red
+        self.green = green
+        self.blue = blue
+        self.opacity = opacity
+    }
+
+    func encode(to encoder: Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
-        let nativeColor = NativeColor(self)
-
-        var red: CGFloat = 0
-        var green: CGFloat = 0
-        var blue: CGFloat = 0
-        var alpha: CGFloat = 0
-        nativeColor.getRed(&red, green: &green, blue: &blue, alpha: &alpha)
-
         try container.encode(red, forKey: .red)
         try container.encode(green, forKey: .green)
         try container.encode(blue, forKey: .blue)
-        try container.encode(alpha, forKey: .opacity)
+        try container.encode(opacity, forKey: .opacity)
     }
 
-    public init(from decoder: Decoder) throws {
+    init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
-        let red = try container.decode(CGFloat.self, forKey: .red)
-        let green = try container.decode(CGFloat.self, forKey: .green)
-        let blue = try container.decode(CGFloat.self, forKey: .blue)
-        let opacity = try container.decode(CGFloat.self, forKey: .opacity)
-
-        self.init(NativeColor(red: red, green: green, blue: blue, alpha: opacity))
+        red = try container.decode(CGFloat.self, forKey: .red)
+        green = try container.decode(CGFloat.self, forKey: .green)
+        blue = try container.decode(CGFloat.self, forKey: .blue)
+        opacity = try container.decode(CGFloat.self, forKey: .opacity)
     }
 }