+++ Sources/DCSettings/Settings/DCStoredValue.swift
@@ -40,7 +40,7 @@ public struct DCStoredValue<ValueType>: DynamicProperty where ValueType: Equatab
     /// Initializes a new `DCStoredValue` instance with the specified key and settings manager.
     ///
     /// This initializer creates a new instance of `DCStoredValue` with the specified key and settings manager.
-    /// The key and default value are required, while the settings manager is optional and defaults to the `.shared` singleton instance.
+    /// The key is required, while the settings manager is optional and defaults to the `.shared` singleton instance.
     ///
     /// If a `DCSetting` instance with the specified key already exists in the settings manager, it will be used to initialize the `StateObject` property.
     /// Otherwise, a runtime error will occur.